Do not migrate livemode to sqlite-in-do
